# perf(expr): add initial microbenchmark for expressions