--- a/src/expr/src/vector_op/agg/mod.rs
+++ b/src/expr/src/vector_op/agg/mod.rs
@@ -23,5 +23,5 @@ mod general_distinct_agg;
 mod general_sorted_grouper;
 mod string_agg;
 
-pub use aggregator::{AggStateFactory, BoxedAggState};
+pub use aggregator::{create_agg_state_unary, AggStateFactory, BoxedAggState};
 pub use general_sorted_grouper::{create_sorted_grouper, BoxedSortedGrouper, EqGroups};