@BugenZhao 2022-12-12
> for primitive types, we can first apply the operation on all values without caring about null values or overflow. The benefits from SIMD are much more than saving computations from null values or errors

I was once thinking this can be done by the compiler. 🥵 However, considering the control flow (side effect) introduced by `checked`, it sounds reasonable for not being optimized.